pyproject.toml
[project]
name = "smart-invoice-saas"
version = "2.0.0"
description = "Smart Invoice SaaS - Automated invoice follow-up with FastAPI backend and Make.com automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Legacy CLI dependencies (kept for porting core logic)
    "google-api-python-client>=2.151.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "APScheduler>=3.10.4",
    "tabulate>=0.9.0",
    "rich>=13.7.0",
    # FastAPI backend dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.0",
    # Auth0
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.3.0",
    # Stripe
    "stripe>=7.0.0",
    # Email
    "sendgrid>=6.11.0",
    "jinja2>=3.1.0",
    # HTTP client
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
